refactor: move go arithmetic filter into language
